move `discord` module to `/lib`
