+++ lib/discord/mod.ts
