+++ lib/discord/register.ts
