+++ servers/lc_dailies/server.ts
@@ -1,5 +1,5 @@
+import * as discord from "../../lib/discord/mod.ts";
 import * as app from "../../app/mod.ts";
-import * as discord from "../../discord/mod.ts";
 import * as lc from "../../lc/mod.ts";
 import * as leaderboard from "../../leaderboard/mod.ts";
 import * as server from "../../lib/server/mod.ts";
